# Faster Write implementation using cpblk