… src/Microsoft.AspNetCore.Server.Kestrel/Internal/Http/FrameHeaders.Generated.cs
… test/Microsoft.AspNetCore.Server.Kestrel.Performance/WritableBufferExtensions.cs